moved load_plotly
- put load_plotly funciton inside plotly
